Historia de usuarios 1 y 2

+++ TurnosRotativosTempl/src/main/java/com/neoris/turnosrotativos/dtos/EmpleadoDTO.java
@@ -5,11 +5,15 @@ import lombok.Setter;
 
 import java.time.LocalDate;
 
+/**
+ * Esta clase se utiliza solo para mostrar información
+ * por lo que no hacen falta verificaciones en ella
+ * */
 @Getter
 @Setter
 public class EmpleadoDTO {
     private Long id;
-    private int nroDocumento;
+    private Long nroDocumento;
     private String nombre;
     private String apellido;
     private String email;
